Clinton conglomerates
i forgot. so it's strange. Mrs Clinton did not say anything to a conglomerate, so the Clinton financial group of oneself said that there was no reason for killing me.
If this keeps going as it is, Article 9 amendment. Japanese politicians are making political contributions to Clinton financial group. Japan is shielded by Russia. and Japan force attack to north Korea under Princess Diana in heaven. Japan is a  facto military power country nationalism. For this reason Japan discards democracy. And we become border guards. The Abe 's Korean drama is already annoying to us which the system is already finished.  so we would be become an orphan of the world.
what a this story is! this is jet-black sunrise.this is rumor.
Aso and parliamentarians sent money to Clinton conglomerates with political contributions as  Akishinomiya will do emperor.  because the child of Princess Diana is convenient. Britain also has only the eldest son to bear the name of the royal family pretty strict with precepts. If I say it in Japan it is the prince of the eldest son. i don't understand Mr. Aso's plan. there is no difference in that Hironomiya is the eldest son of Michiko.  Aso is been so worring by hironomiya of International Convention decided. because Aso would considered that Japanese could goes to the north south?  
However, if you we force attack north, ballistic missiles will fly and abductee victims will have dead people. · We are meaningless north attack that we hope only to back to Japan all abductee victims.
it is been not a joke with Aso's rumor.this is rumor.
Aso and parliamentarians sent money to Clinton conglomerates with political contributions as  Akishinomiya will do emperor.  because the child of Princess Diana is convenient. Britain also has only the eldest son to bear the name of the royal family pretty strict with precepts. If I say it in Japan it is the prince of the eldest son. i don't understand Mr. Aso's plan. there is no difference in that Hironomiya is the eldest son of Michiko.  Aso is been so worring by hironomiya of International Convention decided. because Aso would considered that Japanese could goes to the north south?  
However, if you we force attack north, ballistic missiles will fly and abductee victims will have dead people. · We are meaningless north attack that we hope only to back to Japan all abductee victims.
it is been not a joke with Aso's rumor.